requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
Confluence-py==1.0.1
